+++ CHANGELOG.md
@@ -8,9 +8,19 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `sizeBasedOn` prop to the `Logo` component.
+- Added the `autoComplete` prop to the `TextInput` component for setting the "autocomplete" attribute manually.
+
+### Updates
+
+- Updates max version for Chakra UI to 1.8.9 for minor bug fixes.
+
 ## Fixes
 
 - Adds a z-index on hover to the `SearchBar`'s select icon so it no longer disappears.
+
 
 ## 2.1.1 (October 26, 2023)